Update value type from i32 to i64 in order prevent early overflow

+++ my_iced_project/src/main.rs
@@ -5,7 +5,7 @@ pub struct Counter {
     increment_button: button::State,
     decrement_button: button::State,
     square_button: button::State,
-    value: i32,
+    value: i64,
 }
 
 #[derive(Debug, Clone, Copy)]
